abstracted plexos utilities and made project gid class for gid supersets
