+++ .gitignore
@@ -4,6 +4,10 @@ __pycache__/
 *.py[cod]
 *$py.class
 
+# vim
+tags
+*.swp
+
 # Distribution / packaging
 .Python
 env/
