+++ reVX/plexos/rev_reeds_plexos.py
@@ -14,7 +14,10 @@ from warnings import warn
 import logging
 
 from reV.utilities.execution import SpawnProcessPool
+from reV.plexos.utilities import parse_table_name
 from reVX.handlers.outputs import Outputs
+from reVX.plexos.utilities import DataCleaner, get_coord_labels
+
 
 logger = logging.getLogger(__name__)
 
@@ -496,7 +499,7 @@ class PlexosAggregation:
             warn(wmsg)
             logger.warning(wmsg)
 
-            clabels = self._get_coord_labels(self._sc_build)
+            clabels = get_coord_labels(self._sc_build)
             good_tree = cKDTree(self._sc_build.loc[good_bool, clabels])
             _, i = good_tree.query(self._sc_build.loc[bad_bool, clabels])
 
@@ -549,30 +552,6 @@ class PlexosAggregation:
         output = np.zeros(shape, dtype=np.float32)
         return output
 
-    @staticmethod
-    def _get_coord_labels(df):
-        """Retrieve the coordinate labels from df.
-
-        Parameters
-        ----------
-        df : pd.DataFrame
-            DataFrame with each row being a geo location and two columns
-            containing coordinate labels.
-
-        Returns
-        -------
-        df_coord_labels : list | None
-            Two entry list if coordinate labels are found:
-            ['lat', 'lon'] or ['latitude', 'longitude']
-        """
-
-        df_coord_labels = None
-        if 'lat' in df and 'lon' in df:
-            df_coord_labels = ['lat', 'lon']
-        elif 'latitude' in df and 'longitude' in df:
-            df_coord_labels = ['latitude', 'longitude']
-        return df_coord_labels
-
     def _parse_rev_reeds(self, rev_sc, reeds_build):
         """Parse and combine reV SC and REEDS buildout tables into single table
 
@@ -613,7 +592,8 @@ class PlexosAggregation:
 
         return table
 
-    def _check_rev_reeds_coordinates(self, rev_sc, reeds_build, rev_join_on,
+    @staticmethod
+    def _check_rev_reeds_coordinates(rev_sc, reeds_build, rev_join_on,
                                      reeds_join_on, threshold=0.5):
         """Check that the coordinates are the same in rev and reeds buildouts.
 
@@ -639,8 +619,8 @@ class PlexosAggregation:
             Same as input but without lat/lon columns if matched.
         """
 
-        rev_coord_labels = self._get_coord_labels(rev_sc)
-        reeds_coord_labels = self._get_coord_labels(reeds_build)
+        rev_coord_labels = get_coord_labels(rev_sc)
+        reeds_coord_labels = get_coord_labels(reeds_build)
 
         if rev_coord_labels is not None and reeds_coord_labels is not None:
             reeds_build = reeds_build.sort_values(reeds_join_on)
@@ -685,7 +665,7 @@ class PlexosAggregation:
             with Outputs(self._forecast_fpath) as out:
                 meta_fo = out.meta
 
-            clabels = self._get_coord_labels(meta_cf)
+            clabels = get_coord_labels(meta_cf)
             tree = cKDTree(meta_fo[clabels])
             d, fmap = tree.query(meta_cf[clabels])
             logger.info('Distance (min / mean / max) from generation pixels '
@@ -710,8 +690,8 @@ class PlexosAggregation:
             requested. Values are the Plexos node index.
         """
 
-        plexos_coord_labels = self._get_coord_labels(self._plexos_nodes)
-        sc_coord_labels = self._get_coord_labels(self._sc_build)
+        plexos_coord_labels = get_coord_labels(self._plexos_nodes)
+        sc_coord_labels = get_coord_labels(self._sc_build)
         tree = cKDTree(self._plexos_nodes[plexos_coord_labels])
         d, i = tree.query(self._sc_build[sc_coord_labels], k=k)
         logger.info('Plexos Node KDTree distance min / mean / max: '
@@ -864,272 +844,6 @@ class PlexosAggregation:
         return pa.plexos_meta, pa.time_index, profiles
 
 
-class DataCleaner:
-    """Class for custom Plexos data cleaning procedures."""
-
-    REEDS_NAME_MAP = {'capacity_reV': 'built_capacity',
-                      'capacity_rev': 'built_capacity',
-                      'year': 'reeds_year',
-                      'Year': 'reeds_year'}
-
-    REV_NAME_MAP = {'sq_km': 'area_sq_km',
-                    'capacity': 'potential_capacity',
-                    'resource_ids': 'res_gids',
-                    'resource_ids_cnts': 'gid_counts'}
-
-    PLEXOS_META_COLS = ('gid', 'plexos_id', 'latitude', 'longitude',
-                        'voltage', 'interconnect', 'built_capacity')
-
-    def __init__(self, plexos_meta, profiles):
-        """
-        Parameters
-        ----------
-        plexos_meta : pd.DataFrame
-            Plexos meta data including the built capacity at each plexos node.
-        profiles : np.ndarray
-            2D timeseries array of generation profiles. Number of columns must
-            match the length of the meta data.
-        """
-        if profiles.shape[1] != len(plexos_meta):
-            raise ValueError('Plexos profiles shape does not match meta.')
-        self._plexos_meta = plexos_meta
-        self._profiles = profiles
-
-    @staticmethod
-    def rename_cols(df, name_map):
-        """Do a column rename to make the merge with rev less confusing
-
-        Parameters
-        ----------
-        df : pd.DataFrame
-            Input df with bad or inconsistent column names.
-
-        Parameters
-        ----------
-        df : pd.DataFrame
-            Same as inputs but with better col names.
-        """
-        df = df.rename(columns=name_map)
-        return df
-
-    @staticmethod
-    def reduce_df(df, cols):
-        """Reduce a df to just certain columns.
-
-        Parameters
-        ----------
-        df : pd.DataFrame
-            Dataframe to reduce.
-        cols : list | tuple
-            List of column names to keep.
-
-        Returns
-        -------
-        df : pd.DataFrame
-            Dataframe with only cols if the input df had all cols.
-        """
-
-        cols = [c for c in cols if c in df]
-        return df[cols]
-
-    @staticmethod
-    def pre_filter_plexos_meta(plexos_meta):
-        """Pre-filter the plexos meta data to drop bad node names and
-        duplicate lat/lons.
-
-        Parameters
-        ----------
-        meta_final : pd.DataFrame
-            Plexos meta data.
-
-        Returns
-        -------
-        meta_final : pd.DataFrame
-            Filtered plexos meta data.
-        """
-
-        # as of 8/2019 there were two erroneous plexos nodes with bad names
-        mask = (plexos_meta['plexos_id'] != '#NAME?')
-        plexos_meta = plexos_meta[mask]
-
-        # Several plexos nodes share the same location. As of 8/2019
-        # Josh Novacheck suggests that the duplicate locations can be dropped.
-        plexos_meta = plexos_meta.sort_values(by='voltage', ascending=False)
-        plexos_meta = plexos_meta.drop_duplicates(
-            subset=['latitude', 'longitude'], keep='first')
-        plexos_meta = plexos_meta.sort_values(by='gid')
-
-        return plexos_meta
-
-    @staticmethod
-    def _merge_plexos_meta(meta_final, meta_orig, i_final, i_orig):
-        """Ammend the plexos meta dataframe with data about resource buildouts.
-
-        Parameters
-        ----------
-        meta_final : pd.DataFrame
-            Plexos meta data for the final set of nodes.
-        meta_orig : pd.DataFrame
-            Plexos meta data for the original pre-merge set of nodes.
-        i_final : int
-            Index location (iloc) of the persistent meta data row in
-            meta_final.
-        i_orig : int
-            Index location (iloc) of the meta data row to be merged in
-            meta_orig.
-
-        Returns
-        -------
-        meta_final : pd.DataFrame
-            Plexos meta data for the final set of nodes.
-        """
-
-        i_final = meta_final.index.values[i_final]
-        i_orig = meta_orig.index.values[i_orig]
-
-        cols = ['res_gids', 'gen_gids', 'res_built', 'built_capacity']
-
-        for col in cols:
-            val_final = meta_final.loc[i_final, col]
-            val_orig = meta_orig.loc[i_orig, col]
-            if not isinstance(val_final, type(val_orig)):
-                raise TypeError('Mismatch in column dtype for plexos meta!')
-            if isinstance(val_final, str):
-                val_final = json.loads(val_final)
-                val_orig = json.loads(val_orig)
-                val_final += val_orig
-                val_final = str(val_final)
-            else:
-                val_final += val_orig
-
-            meta_final.loc[i_final, col] = val_final
-
-        return meta_final
-
-    def merge_small(self, capacity_threshold=20.0):
-        """Merge small plexos buildout nodes into closest bigger nodes.
-
-        Parameters
-        ----------
-        capacity_threshold : float
-            Capacity threshold, nodes with built capacities less than this
-            will be merged into bigger nodes.
-
-        Returns
-        -------
-        meta : pd.DataFrame
-            New plexos node meta data with updated built capacities.
-        profiles : np.ndarray
-            New profiles with big nodes having absorbed additional generation
-            from bigger nodes.
-        """
-
-        small = (self._plexos_meta['built_capacity'] < capacity_threshold)
-        big = (self._plexos_meta['built_capacity'] >= capacity_threshold)
-
-        n_nodes = np.sum(big)
-        if (n_nodes == len(self._plexos_meta) or n_nodes == 0):
-            meta = None
-            profiles = None
-
-        else:
-            meta = self._plexos_meta[big]
-            profiles = self._profiles[:, big.values]
-            logger.info('Merging plexos nodes from {} to {} due to small '
-                        'nodes.'.format(len(self._plexos_meta), len(meta)))
-
-            labels = PlexosAggregation._get_coord_labels(self._plexos_meta)
-            tree = cKDTree(meta[labels])
-            _, nn_ind = tree.query(self._plexos_meta[labels], k=len(meta))
-
-            for i in range(len(self._plexos_meta)):
-                if small.values[i]:
-                    for nn in nn_ind[i, :]:
-                        if big.values[nn]:
-                            meta = self._merge_plexos_meta(meta,
-                                                           self._plexos_meta,
-                                                           nn, i)
-                            profiles[:, nn] += self._profiles[:, i]
-
-                            break
-
-        return meta, profiles
-
-    def merge_extent(self, new_meta, new_profiles):
-        """Merge a new set of plexos node aggregation data into the self attr.
-
-        Parameters
-        ----------
-        new_meta : pd.DataFrame
-            A new set of Plexos node meta data to be merged into the meta in
-            self.
-        new_profiles : np.ndarray
-            A new set of plexos node profiles corresponding to new_meta to be
-            merged into the profiles in self where the meta data overlaps with
-            common nodes.
-        """
-
-        keep_index = []
-
-        logger.info('Merging extents with {} and {} nodes ({} total).'
-                    .format(len(self._plexos_meta), len(new_meta),
-                            len(self._plexos_meta) + len(new_meta)))
-
-        for i, ind in enumerate(new_meta.index.values):
-            lookup = (self._plexos_meta.gid.values == new_meta.loc[ind, 'gid'])
-            if any(lookup):
-                i_self = np.where(lookup)[0]
-                if len(i_self) > 1:
-                    warn('Duplicate PLEXOS node GIDs in base plexos meta!')
-                else:
-                    i_self = i_self[0]
-
-                logger.debug('Merging plexos node IDs {} and {} '
-                             '(gids {} and {})'.format(
-                                 self._plexos_meta.iloc[i_self]['plexos_id'],
-                                 new_meta.iloc[i]['plexos_id'],
-                                 self._plexos_meta.iloc[i_self]['gid'],
-                                 new_meta.iloc[i]['gid']))
-
-                self._merge_plexos_meta(self._plexos_meta, new_meta, i_self, i)
-                self._profiles[:, i_self] += new_profiles[:, i]
-            else:
-                keep_index.append(i)
-
-        new_meta = new_meta.loc[new_meta.index.values[keep_index]]
-        new_profiles = new_profiles[:, keep_index]
-
-        self._plexos_meta = pd.concat([self._plexos_meta, new_meta], axis=0,
-                                      ignore_index=True)
-        self._profiles = np.hstack((self._profiles, new_profiles))
-
-        logger.info('Merged extents. Output has {} nodes.'
-                    .format(len(self._plexos_meta)))
-
-    def merge_multiple_extents(self, meta_list, profile_list):
-        """Merge multiple plexos extents into the self attrs.
-
-        Parameters
-        ----------
-        meta_list : list
-            List of new meta data extents to merge into self.
-        profile_list : list
-            List of new gen profile to merge into self.
-
-        Returns
-        -------
-        meta : pd.DataFrame
-            Merged plexos node meta data.
-        profiles : np.ndarray
-            New profiles with merged profiles for matching nodes.
-        """
-
-        for i, meta in enumerate(meta_list):
-            self.merge_extent(meta, profile_list[i])
-
-        return self._plexos_meta, self._profiles
-
-
 class Manager:
     """Plexos job manager."""
 
@@ -1162,18 +876,24 @@ class Manager:
         db_port : int
             Database port.
         """
-        self.plexos_nodes = self._parse_name(plexos_nodes, wait=wait,
-                                             db_host=db_host, db_user=db_user,
-                                             db_pass=db_pass, db_port=db_port)
+        self.plexos_nodes = parse_table_name(plexos_nodes, wait=wait,
+                                             db_host=db_host,
+                                             db_user=db_user,
+                                             db_pass=db_pass,
+                                             db_port=db_port)
         self.plexos_nodes = DataCleaner.pre_filter_plexos_meta(
             self.plexos_nodes)
 
-        self.rev_sc = self._parse_name(rev_sc, wait=wait,
-                                       db_host=db_host, db_user=db_user,
-                                       db_pass=db_pass, db_port=db_port)
-        self.reeds_build = self._parse_name(reeds_build, wait=wait,
-                                            db_host=db_host, db_user=db_user,
-                                            db_pass=db_pass, db_port=db_port)
+        self.rev_sc = parse_table_name(rev_sc, wait=wait,
+                                       db_host=db_host,
+                                       db_user=db_user,
+                                       db_pass=db_pass,
+                                       db_port=db_port)
+        self.reeds_build = parse_table_name(reeds_build, wait=wait,
+                                            db_host=db_host,
+                                            db_user=db_user,
+                                            db_pass=db_pass,
+                                            db_port=db_port)
 
         self.rev_sc = DataCleaner.rename_cols(self.rev_sc,
                                               DataCleaner.REV_NAME_MAP)
@@ -1190,52 +910,6 @@ class Manager:
             if not os.path.exists(self.forecast_fpath):
                 raise FileNotFoundError('Could not find forecast_fpath: {}'
                                         .format(forecast_fpath))
-
-    @staticmethod
-    def _parse_name(name, wait=300, db_host='gds_edit.nrel.gov',
-                    db_user=None, db_pass=None, db_port=5432):
-        """Parse a dataframe from an input name.
-
-        Parameters
-        ----------
-        name : str | pd.DataFrame
-            CSV file path or database.schema.name
-        wait : int
-            Integer seconds to wait for DB connection to become available
-            before raising exception.
-        db_host : str
-            Database host name.
-        db_user : str
-            Your database user name.
-        db_pass : str
-            Database password (None if your password is cached).
-        db_port : int
-            Database port.
-
-        Returns
-        -------
-        df : pd.DataFrame
-            Extracted table
-        """
-
-        if isinstance(name, str):
-            if name.endswith('.csv'):
-                df = pd.read_csv(name)
-            elif len(name.split('.')) == 3:
-                from reVX.handlers.database import Database
-                db, schema, table = name.split('.')
-                df = Database.get_table(table, schema, db, wait=wait,
-                                        db_host=db_host, db_user=db_user,
-                                        db_pass=db_pass, db_port=db_port)
-
-        elif isinstance(name, pd.DataFrame):
-            df = name
-
-        else:
-            raise TypeError('Could not recognize input table name: '
-                            '{} with type {}'.format(name, type(name)))
-
-        return df
 
     @classmethod
     def main(cls, plexos_nodes, rev_sc, reeds_build, cf_fpath,
